remove main
