--- apiapp/cmd/myapp/main.go
@@ -1,11 +1,0 @@
-package main
-
-//repositoryディレクトリをインポート
-import (
-	"apiapp/repository"
-)
-
-func main() {
-	repo := repository.NewProductRepository()
-	repo.Execute(1)
-}
