--- apiapp/repository/interface.go
@@ -1,5 +1,0 @@
-package repository
-
-type ProductRepository interface {
-	Execute(productId int)
-}
