+++ apiapp/repository/product_repository.go
@@ -1,0 +1,7 @@
+package repository
+
+import "apiapp/product"
+
+type ProductRepository interface {
+	GetProductDetail(productId int) (*product.Product, error)
+}
