+++ apiapp/repository/product_repository_impl.go
@@ -14,16 +14,6 @@ func NewProductRepository() ProductRepository {
 	return &ProductRepositoryImpl{}
 }
 
-func (repo *ProductRepositoryImpl) Execute(productId int) {
-	product, err := repo.GetProductDetail(productId)
-	if err != nil {
-		fmt.Println("Error:", err)
-		return
-	}
-
-	fmt.Printf("ID: %d, Title: %s\n", product.ID, product.Title)
-}
-
 func (repo *ProductRepositoryImpl) GetProductDetail(productId int) (*product.Product, error) {
 	url := fmt.Sprintf("https://dummyjson.com/products/%d", productId)
 	resp, err := http.Get(url)
